refactor(announcements): move AnnouncementRepository interface to application/usecases
Applies the v0.141.0 DIP template to announcements: relocate the single
AnnouncementRepository interface + its co-located AnnouncementFilter
struct from announcements/domain/repositories/ to consumer-side
announcements/application/usecases/. Empty domain/repositories/ dir
deleted per CLAUDE.md DDD-gate.

Mechanical scope (no behavior change, build/test/lint green per package):
- file relocated (git rename detected)
- 3 consumers updated to drop the domain/repositories import:
  - application/usecases/announcement_usecase.go (3 sites)
  - application/usecases/announcement_usecase_test.go (4 mock-method sites)
  - infrastructure/persistence/announcement_repository_pg.go (5 sites;
    aliased import switched to application/usecases)
- compile-time guarantee added in pg repo:
  var _ usecases.AnnouncementRepository = (*AnnouncementRepositoryPG)(nil)
  (mirror to the v0.141.0 users module precedent)

Unblocks handler-mocking against the narrow port; coverage backfill for
the 0% PG repo lands in the next commit (Phase 6 #196 sprint).
